web/requirements.txt
gunicorn
gevent
flask_sqlalchemy
flask_bcrypt
flask_wtf
pytz
pydub
azure-storage-blob
hume
opencc
flask_migrate
zhconv
email-validator
pandas